refactor(utils/misc): move color utils into separate file & add docstring for #5187
